Comment out all dead & offending code
Fixes annotation processing problems & startup exceptions.

All dead code is accompanied with the TODOs for later cleanup.
