+++ lyo-services/domain-pddl/src/main/java/eu/scott/warehouse/lib/hazelcast/HCData.kt
@@ -1,8 +1,6 @@
 package eu.scott.warehouse.lib.hazelcast
 
-import java.io.Serializable
-import java.util.*
-
+/*
 @Deprecated("Phasing Hazelcast out of the SCOTT sandbox")
 data class HCData(val url: String, val uuid: String, val headers: Array<String>) :
     Serializable {
@@ -24,3 +22,4 @@ data class HCData(val url: String, val uuid: String, val headers: Array<String>)
         return result
     }
 }
+*/
